# docs/swagger.yaml
definitions:
...
  response.Response:
    properties:
      body: {}
      code:
        type: integer
      message:
        type: string
    type: object
  user.changePasswordRequest:
    properties:
      password:
        example: password
        maxLength: 32
        minLength: 6
        type: string
    required:
    - password
    type: object
  user.changeUserRoleRequest:
    properties:
      role:
        example: 0
        type: integer
      userID:
        example: "21"
        type: string
    required:
    - role
    - userID
    type: object
  user.createUserRequest:
    properties:
      email:
        example: example@gmail.com
        type: string
      password:
        example: password
        maxLength: 32
        minLength: 6
        type: string
      phone:
        type: string
      username:
        example: John Shnow
        maxLength: 50
        minLength: 2
        type: string
    required:
    - email
    - password
    - phone
    - username
    type: object
host: localhost:8080
info:
  contact:
    email: dragon090986@gmail.com
    name: Kirill Shaforostov
  description: API for recipe recommendation application
  title: Recipe API
  version: 0.0.1
paths:
  /admin/users/role/{id}:
    patch:
      consumes:
      - application/json
      description: Change the role of a user with the given ID
      parameters:
      - description: User information
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/user.changeUserRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BasicAuth: []
      summary: Change user role
      tags:
      - User
...
  /users:
    delete:
      consumes:
      - application/json
      description: Delete the currently logged in user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BasicAuth: []
      summary: Delete user
      tags:
      - User
    patch:
      consumes:
      - application/json
      description: Update information of the currently logged in user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BasicAuthcAuth: []
      summary: Update user information
      tags:
      - User
    post:
      consumes:
      - application/json
      description: Create a new user with the given information
      parameters:
      - description: User information
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/user.createUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Create a new user
      tags:
      - User
  /users/{userID}:
    get:
      consumes:
      - application/json
      description: Get information about the currently logged in user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BasicAuth: []
      summary: Get user information
      tags:
      - User
  /users/password:
    patch:
      consumes:
      - application/json
      description: Change the password of the currently logged in user
      parameters:
      - description: User information
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/user.changePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BasicAuthcAuth: []
      summary: Change user password
      tags:
      - User
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"